+++ RibosoftAlgo/test/test_fold.cpp
@@ -7,7 +7,7 @@ using namespace ribosoft;
 
 TEST_CASE("default", "[fold]") {
     fold_output* output = nullptr;
-    int size;
+    size_t size;
     R_STATUS status = fold("AUGUCUUAGGUGAUACGUGC", output, size);
     REQUIRE(status == R_SUCCESS::R_STATUS_OK);
     REQUIRE(strcmp(output[0].structure, ".((((......)))).....") == 0);
@@ -19,7 +19,7 @@ TEST_CASE("default", "[fold]") {
 
 TEST_CASE("invalid sequence", "[fold]") {
     fold_output* output = nullptr;
-    int size;
+    size_t size;
     R_STATUS status = fold("wfef", output, size);
     REQUIRE(status == R_APPLICATION_ERROR::R_INVALID_NUCLEOTIDE);
 }
